+++ plugin/app/Services/NotificationService.php
@@ -4,9 +4,7 @@ namespace TTU\Charon\Services;
 
 use Illuminate\Support\Facades\Log;
 use stdClass;
-use TTU\Charon\Models\Charon;
-use TTU\Charon\Repositories\SubmissionsRepository;
-use Zeizig\Moodle\Globals\User;
+use Zeizig\Moodle\Models\User;
 
 /**
  * Class NotificationService.
@@ -15,72 +13,11 @@ use Zeizig\Moodle\Globals\User;
  */
 class NotificationService
 {
-    /**
-     * @var SubmissionsRepository
-     */
-    private $submissionsRepository;
-
-    /**
-     * NotificationService constructor.
-     */
-    public function __construct(SubmissionsRepository $submissionsRepository)
-    {
-        $this->submissionsRepository = $submissionsRepository;
-    }
-
-    /**
-     * Send a notification to student about a new comment submission.
-     *
-     * @param int $submissionId
-     * @param string $messageText
-     * @param Charon $charon
-     * @param string $filePath
-     */
-    public function sendNotificationToStudent(
-        int $submissionId,
-        string $messageText,
-        Charon $charon,
-        string $filePath
-    ) {
-        $teacher = app(User::class)->currentUser();
-
-        $submission = $this->submissionsRepository->find($submissionId);
-
-        $students = $this->submissionsRepository->findAllUsersAssociated($submissionId);
-
-        $cm_id = $charon->courseModule()->id;
-        $url = '/mod/charon/view.php?id=' . $cm_id;
-
-        $messageText = htmlspecialchars($messageText);
-        $messageText = str_replace( "\n", '<br />', $messageText );
-
-        $messageTextHtml = <<<EOT
-<h4>$charon->name</h4><br>
-<b>You've got a new comment for the submission that was submitted at 
-$submission->created_at</b><br>
-<b>Author: $teacher->firstname $teacher->lastname</b><br>
-<b>File that was commented: $filePath</b><br><br>
-<p style="white-space: pre-wrap">$messageText</p>
-EOT;
-
-        foreach ($students as $student) {
-            $this->sendNotification(
-                $teacher,
-                $student,
-                'comment',
-                "New comment: " . $charon->name,
-                $messageText,
-                $messageTextHtml,
-                $url,
-                $charon->name
-            );
-        }
-    }
 
     /**
      * Send a notification to user.
      *
-     * @param \Zeizig\Moodle\Models\User $userFrom
+     * @param User $userFrom
      * @param stdClass $userTo
      * @param string $notificationName
      * @param string $notificationSubject
@@ -89,8 +26,8 @@ EOT;
      * @param string $urlToDirectTo
      * @param string $urlNameToDirect
      */
-    private function sendNotification(
-        \Zeizig\Moodle\Models\User $userFrom,
+    public function sendNotification(
+        User $userFrom,
         stdClass $userTo,
         string $notificationName,
         string $notificationSubject,